+++ packages/template-retail-react-app/app/components/_app/index.jsx
@@ -49,6 +49,7 @@ import AboveHeader from '@salesforce/retail-react-app/app/components/_app/partia
 import StoreLocatorModal from '@salesforce/retail-react-app/app/components/store-locator-modal'
 // Hooks
 import {AuthModal, useAuthModal} from '@salesforce/retail-react-app/app/hooks/use-auth-modal'
+import {DntModal, useDntModal} from '@salesforce/retail-react-app/app/hooks/use-dnt-modal'
 import {AddToCartModalProvider} from '@salesforce/retail-react-app/app/hooks/use-add-to-cart-modal'
 import useMultiSite from '@salesforce/retail-react-app/app/hooks/use-multi-site'
 import {useCurrentCustomer} from '@salesforce/retail-react-app/app/hooks/use-current-customer'
@@ -129,6 +130,7 @@ const App = (props) => {
     const history = useHistory()
     const location = useLocation()
     const authModal = useAuthModal()
+    const dntModal = useDntModal()
     const {site, locale, buildUrl} = useMultiSite()
 
     const [isOnline, setIsOnline] = useState(true)
@@ -425,6 +427,7 @@ const App = (props) => {
                                 {!isCheckout ? <Footer /> : <CheckoutFooter />}
 
                                 <AuthModal {...authModal} />
+                                <DntModal {...dntModal} />
                             </AddToCartModalProvider>
                         </Box>
                     </CurrencyProvider>